refactor: create reusable action header component for settings form pages
